+++ src/Hy3Layout.hpp
@@ -20,6 +20,12 @@ enum class ShiftDirection {
 	Down,
 	Right,
 };
+inline static constexpr char getShiftDirectionChar(ShiftDirection direction) {
+	return direction == ShiftDirection::Left ? 'l'
+	     : direction == ShiftDirection::Up   ? 'u'
+	     : direction == ShiftDirection::Down ? 'd'
+	                                         : 'r';
+}
 
 enum class Axis { None, Horizontal, Vertical };
 
@@ -113,6 +119,10 @@ public:
 	void shiftNode(Hy3Node&, ShiftDirection, bool once, bool visible);
 	void shiftWindow(const PHLWORKSPACE& workspace, ShiftDirection, bool once, bool visible);
 	void shiftFocus(const PHLWORKSPACE& workspace, ShiftDirection, bool visible, bool warp);
+	void toggleFocusLayer(const PHLWORKSPACE& workspace, bool warp);
+	bool shiftMonitor(Hy3Node&, ShiftDirection, bool follow);
+	Hy3Node* focusMonitor(ShiftDirection);
+
 	void warpCursor();
 	void moveNodeToWorkspace(const PHLWORKSPACE& origin, std::string wsname, bool follow);
 	void changeFocus(const PHLWORKSPACE& workspace, FocusShift);
@@ -129,6 +139,8 @@ public:
 	static void warpCursorToBox(const Vector2D& pos, const Vector2D& size);
 
 	bool shouldRenderSelected(const PHLWINDOW&);
+	PHLWINDOW findTiledWindowCandidate(const PHLWINDOW& from);
+	PHLWINDOW findFloatingWindowCandidate(const PHLWINDOW& from);
 
 	Hy3Node* getWorkspaceRootGroup(const PHLWORKSPACE& workspace);
 	Hy3Node* getWorkspaceFocusedNode(